+++ RandomUsers/RandomUsers.Android/MainActivity.cs
@@ -26,12 +26,13 @@ namespace RandomUsers.Droid
             ToolbarResource = Resource.Layout.Toolbar;
 
             base.OnCreate(bundle);
+            
             global::Xamarin.Forms.Forms.Init(this, bundle);
             ImageCircle.Forms.Plugin.Droid.ImageCircleRenderer.Init();
             _sensorManager = (SensorManager)GetSystemService(Context.SensorService);
             _sensor = _sensorManager.GetDefaultSensor(SensorType.Accelerometer);
 
-            _shakeDetector = new ShakeDetector();
+            
 #if DEBUG
             StreamReader strm = new StreamReader(Assets.Open("data.json"));
             var json = strm.ReadToEnd();
@@ -41,7 +42,7 @@ namespace RandomUsers.Droid
             LoadApplication(new App(json));
 #else
             var app = new App();
-
+            _shakeDetector = new ShakeDetector();
             _shakeDetector.Shaked += (sender, shakeCount) =>
             {
                 lock (_syncLock)
@@ -49,21 +50,27 @@ namespace RandomUsers.Droid
                     app.Shake(shakeCount);
                 }
             };
-
+            
             LoadApplication(app);
 #endif
         }
-
+        
         protected override void OnResume()
         {
             base.OnResume();
-            _sensorManager.RegisterListener(_shakeDetector, _sensor, SensorDelay.Ui);
+            if (_shakeDetector != null)
+            {
+                _sensorManager.RegisterListener(_shakeDetector, _sensor, SensorDelay.Ui);
+            }
         }
 
         protected override void OnPause()
         {
             base.OnPause();
-            _sensorManager.UnregisterListener(_shakeDetector);
+            if (_shakeDetector != null)
+            {
+                _sensorManager.UnregisterListener(_shakeDetector);
+            }
         }
     }
 }
