+++ RandomUsers/RandomUsers/FilterPage.xaml.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using Xamarin.Forms;
+using Xamarin.Forms.Xaml;
+
+namespace RandomUsers
+{
+    [XamlCompilation(XamlCompilationOptions.Compile)]
+    public partial class FilterPage : ContentPage
+    {
+        public FilterPage()
+        {
+            InitializeComponent();
+        }
+    }
+}
