+++ RandomUsers/RandomUsers/Models/ResultInfo.cs
@@ -1,0 +1,10 @@
+﻿namespace RandomUsers.Models
+{
+    public class ResultInfo
+    {
+        public string Seed { get; set; }
+        public int Results { get; set; }
+        public int Page { get; set; }
+        public string Version { get; set; }
+    }
+}
