+++ RandomUsers/RandomUsers/Models/UserLocation.cs
@@ -1,0 +1,11 @@
+﻿namespace RandomUsers.Models
+{
+
+    public class UserLocation
+    {
+        public string Street { get; set; }
+        public string City { get; set; }
+        public string State { get; set; }
+        public string PostCode { get; set; }
+    }
+}
