+++ RandomUsers/RandomUsers/ViewModel/PageViewModelBase.cs
@@ -6,11 +6,18 @@ namespace RandomUsers.ViewModel
 
     public class PageViewModelBase : ViewModelBase
     {
+        public App App
+        {
+            get
+            {
+                return (App)Application.Current;
+            }
+        }
         protected INavigation Navigation
         {
             get
             {
-                return Application.Current.MainPage.Navigation;
+                return (Application.Current as App).MainPage.Navigation;
             }
         }
     }
