+++ RandomUsers/RandomUsers/ViewModel/ViewModelLocator.cs
@@ -27,7 +27,7 @@ namespace RandomUsers
                 return ServiceLocator.Current.GetInstance<MainPageViewModel>();
             }
         }
-
+        
         public UserDetailViewModel Detail
         {
             get
